requirements.txt
fastapi==0.115.5
uvicorn==0.31.0
pydantic==2.12.5
akshare==1.18.7
pandas==2.3.3
numpy
pywebview==6.1
curl_cffi
beautifulsoup4==4.14.3
python-docx==1.1.2
pdfminer.six==20231228
PyPDF2
PyMuPDF
sentence-transformers
scikit-learn
openai
webdriver_manager
passlib[bcrypt]
python-jose[cryptography]
python-multipart
redis==5.0.1
duckduckgo-search

scipy
